+++ src/main/java/roomescape/service/dto/request/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.web.request;
+package roomescape.service.dto.request;
 
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
