+++ src/main/java/roomescape/service/dto/response/ReservationResponse.java
@@ -1,4 +1,4 @@
-package roomescape.web.response;
+package roomescape.service.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
